# Codigo/supportplatform-service/src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
...
			"CommitMetricsDTO": {
				"properties": {
					"totalCommitCount": {
						"type": "number",
						"format": "double"
					},
					"commitsPerContributor": {
						"items": {
							"properties": {
								"commtiPercentage": {
									"type": "number",
									"format": "double"
								},
								"commitCount": {
									"type": "number",
									"format": "double"
								},
								"contribuitor": {
									"properties": {
										"githubAvatarUrl": {
											"type": "string"
										},
										"githubLogin": {
											"type": "string"
										},
										"githubName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"githubAvatarUrl",
										"githubLogin",
										"githubName",
										"id"
									],
									"type": "object"
								}
							},
							"required": [
								"commtiPercentage",
								"commitCount"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"totalCommitCount",
					"commitsPerContributor"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"FileChangeMetricsDTO": {
				"properties": {
					"totalAdditions": {
						"type": "number",
						"format": "double"
					},
					"totalDeletions": {
						"type": "number",
						"format": "double"
					},
					"fileCount": {
						"type": "number",
						"format": "double"
					},
					"fileChangesPerContributor": {
						"items": {
							"properties": {
								"fileCount": {
									"type": "number",
									"format": "double"
								},
								"deletions": {
									"properties": {
										"percentage": {
											"type": "number",
											"format": "double"
										},
										"sum": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"percentage",
										"sum"
									],
									"type": "object"
								},
								"addtions": {
									"properties": {
										"percentage": {
											"type": "number",
											"format": "double"
										},
										"sum": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"percentage",
										"sum"
									],
									"type": "object"
								},
								"contribuitor": {
									"properties": {
										"githubAvatarUrl": {
											"type": "string"
										},
										"githubLogin": {
											"type": "string"
										},
										"githubName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"githubAvatarUrl",
										"githubLogin",
										"githubName",
										"id"
									],
									"type": "object"
								}
							},
							"required": [
								"fileCount",
								"deletions",
								"addtions"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"totalAdditions",
					"totalDeletions",
					"fileCount",
					"fileChangesPerContributor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FileTypeMetricDTO": {
				"properties": {
					"count": {
						"type": "number",
						"format": "double"
					},
					"extension": {
						"type": "string"
					}
				},
				"required": [
					"count",
					"extension"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FileTypeMetricsDTO": {
				"properties": {
					"general": {
						"items": {
							"$ref": "#/components/schemas/FileTypeMetricDTO"
						},
						"type": "array"
					},
					"perContributor": {
						"items": {
							"properties": {
								"fileTypes": {
									"items": {
										"$ref": "#/components/schemas/FileTypeMetricDTO"
									},
									"type": "array"
								},
								"contributor": {
									"properties": {
										"githubAvatarUrl": {
											"type": "string",
											"nullable": true
										},
										"githubLogin": {
											"type": "string"
										},
										"githubName": {
											"type": "string",
											"nullable": true
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"githubAvatarUrl",
										"githubLogin",
										"githubName",
										"id"
									],
									"type": "object"
								}
							},
							"required": [
								"fileTypes"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"general",
					"perContributor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IssueMetricsDTO": {
				"properties": {
					"issueDataPerContributor": {
						"items": {
							"properties": {
								"authoredIssuesCount": {
									"type": "number",
									"format": "double"
								},
								"assignedIssuesCount": {
									"type": "number",
									"format": "double"
								},
								"contributor": {
									"properties": {
										"githubAvatarUrl": {
											"type": "string",
											"nullable": true
										},
										"githubLogin": {
											"type": "string",
											"nullable": true
										},
										"githubName": {
											"type": "string",
											"nullable": true
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"githubAvatarUrl",
										"githubLogin",
										"githubName",
										"id"
									],
									"type": "object"
								}
							},
							"required": [
								"authoredIssuesCount",
								"assignedIssuesCount"
							],
							"type": "object"
						},
						"type": "array"
					},
					"issuesOpennedCount": {
						"type": "number",
						"format": "double"
					},
					"issuesClosedCount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"issueDataPerContributor",
					"issuesOpennedCount",
					"issuesClosedCount"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CommitQualityMetricsDTO": {
				"properties": {
					"generalCommitQualityLevel": {
						"items": {
							"properties": {
								"qualityLevelCount": {
									"type": "number",
									"format": "double"
								},
								"qualityLevel": {
									"type": "number",
									"format": "double"
								}
							},
							"required": [
								"qualityLevelCount",
								"qualityLevel"
							],
							"type": "object"
						},
						"type": "array"
					},
					"commitQualityPerContributor": {
						"items": {
							"properties": {
								"commitQualityLevel": {
									"items": {
										"properties": {
											"qualityLevelCount": {
												"type": "number",
												"format": "double"
											},
											"qualityLevel": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"qualityLevelCount",
											"qualityLevel"
										],
										"type": "object"
									},
									"type": "array"
								},
								"contributor": {
									"properties": {
										"githubAvatarUrl": {
											"type": "string"
										},
										"githubLogin": {
											"type": "string"
										},
										"githubName": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"githubAvatarUrl",
										"githubLogin",
										"githubName",
										"id"
									],
									"type": "object"
								}
							},
							"required": [
								"commitQualityLevel"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"generalCommitQualityLevel",
					"commitQualityPerContributor"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "supportplatform-service",
		"version": "1.0.0",
		"description": "SupportPlatform service for GitHub Project Evaluation Support Platform",
		"license": {
			"name": "CC-BY-4.0"
		},
		"contact": {
			"name": "Lucas Ângelo Oliveira Martins Rocha ",
			"url": "https://lucasangelo.com/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/repository/{repositoryId}/metric/commit": {
			"get": {
				"operationId": "GetCommitMetrics",
				"responses": {
					"200": {
						"description": "Commit metrics",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommitMetricsDTO"
								}
							}
						}
					}
				},
				"tags": [
					"metrics"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "repositoryId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "startedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "branchName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "contributor",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "filterWithNoContributor",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/repository/{repositoryId}/metric/changes": {
			"get": {
				"operationId": "GetChangesMetrics",
				"responses": {
					"200": {
						"description": "File change metrics",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FileChangeMetricsDTO"
								}
							}
						}
					}
				},
				"tags": [
					"metrics"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "repositoryId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "startedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "branchName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "contributor",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "filterWithNoContributor",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/repository/{repositoryId}/metric/file-types": {
			"get": {
				"operationId": "GetFileTypesMetrics",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FileTypeMetricsDTO"
								}
							}
						}
					}
				},
				"tags": [
					"metrics"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "repositoryId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "startedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "branchName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "contributor",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "filterWithNoContributor",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/repository/{repositoryId}/metric/issues": {
			"get": {
				"operationId": "GetIssuesMetrics",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IssueMetricsDTO"
								}
							}
						}
					}
				},
				"tags": [
					"metrics"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "repositoryId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "startedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "contributor",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "filterWithNoContributor",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/repository/{repositoryId}/metric/commit-quality": {
			"get": {
				"operationId": "GetCommitQualityMetrics",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CommitQualityMetricsDTO"
								}
							}
						}
					}
				},
				"tags": [
					"metrics"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "repositoryId",
						"required": true,
						"schema": {
							"format": "int32",
							"type": "integer",
							"minimum": 1
						}
					},
					{
						"in": "query",
						"name": "startedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endedAt",
						"required": false,
						"schema": {
							"format": "date",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "branchName",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "contributor",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "filterWithNoContributor",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}